Навод порядка и продолжение создания виртуальной машины и исполнения байткода

+++ DialogMaker.Core.Scripting/Runtime/Primitives/DSharpBytecodeOperation.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// D# bytecode operation code
     /// </summary>
-    public enum DSharpBytecodeOperation : short
+    public enum DSharpBytecodeOperation : byte
     {
         /// <summary>
         /// Push value to stack
